Revert "LPS-99657 Class that implements ConfigurationBeanDeclaration should be in package *.settings.definition"
This reverts commit 75f63905b1b2324de4af572d410ace0c2bcd3ae5.
